+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/annotator/Protobuf3Annotator.kt
@@ -4,6 +4,7 @@ import com.intellij.lang.annotation.AnnotationHolder
 import com.intellij.lang.annotation.Annotator
 import com.intellij.lang.annotation.HighlightSeverity
 import com.intellij.psi.PsiElement
+import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufBuiltInOptionName
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufExtendDefinition
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufExtensionStatement
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufFieldDefinition
@@ -12,8 +13,8 @@ import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufImportStatement
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufMessageDefinition
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufOneofBody
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufVisitor
+import io.kanro.idea.plugin.protobuf.lang.psi.isFieldDefaultOption
 import io.kanro.idea.plugin.protobuf.lang.psi.primitive.ProtobufElement
-import io.kanro.idea.plugin.protobuf.lang.psi.resolve
 import io.kanro.idea.plugin.protobuf.lang.psi.weak
 import io.kanro.idea.plugin.protobuf.lang.support.Options
 
@@ -83,6 +84,17 @@ class Protobuf3Annotator : Annotator {
                         "'weak' import is not supported in proto3."
                     )
                         .range(o.importLabel?.textRange ?: o.textRange)
+                        .create()
+                }
+            }
+
+            override fun visitBuiltInOptionName(o: ProtobufBuiltInOptionName) {
+                if (o.isFieldDefaultOption()) {
+                    holder.newAnnotation(
+                        HighlightSeverity.ERROR,
+                        "'default' option is not supported in proto3."
+                    )
+                        .range(o.textRange)
                         .create()
                 }
             }