+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/psi/Extension.kt
@@ -163,6 +163,10 @@ fun ProtobufBuiltInOptionName.isFieldDefaultOption(): Boolean {
     return this.textMatches("default") && parentOfType<ProtobufOptionOwner>() is ProtobufFieldDefinition
 }
 
+fun ProtobufBuiltInOptionName.isFieldJsonNameOption(): Boolean {
+    return this.textMatches("json_name") && parentOfType<ProtobufOptionOwner>() is ProtobufFieldDefinition
+}
+
 fun ProtobufTypeName.absolutely(): Boolean {
     return firstChild !is ProtobufSymbolName
 }